Refactor: Aquatic Event loop
